removed Ratliff VHACD since he has not repond in moth,
also fixing few waning error when compiling with all warning (wip)

+++ newton-4.00/applications/ndSandbox/toolbox/ndShaderCache.cpp
@@ -66,7 +66,7 @@ void ndShaderCache::LoadShaderCode (const char* const filename, char* const buff
 	
 	size = ftell (file);
 	fseek (file, 0, SEEK_SET); 
-	size_t error = fread (buffer, size, 1, file);
+	size_t error = fread (buffer, size_t(size), 1, file);
 	// for GCC shit
 	ndAssert (error); error = 0;
 	fclose (file);
